raspberrypi: install all firmware files
Install all firmware files to the target boot partition.

The current solution where only one set of firmware files
(startXXX.elf and fixupXXX.elf) are installed does not work.

The firmware files in the boot partition must not be renamed.
(the current solution installs the files as start.elf and fixup.elf
always)

The names must be preserved since different files will be choosen by
the boot ROM depending on the parameters in config.txt.

Example:
When start_x=1 is set in config.txt, the boot ROM will look for
start_x.elf

The parameters might change during the lifetime of the device so
it is safest to install all files.

Signed-off-by: Erik Stromdahl <erik.stromdahl@gmail.com>